Added proposed method, justification, assumptions, and data processing plan as well as minor changes to previous cells
I added the plan I am going to follow to do the analysis and justified it. I also wrote down assumptions I will make along the way.
I changed the way I named variables in previous cells to make it more uniform. I changed the total_playtime to played_hours as that is what is used in the files.